vision_agent/requirements.txt
fastapi==0.115.12
uvicorn==0.29.0
httpx==0.27.0
pillow==10.2.0
python-multipart==0.0.9
torch==2.2.0
timm==0.9.16
google-generativeai==0.8.5
